separate SL dependant selections from common selection code
