+++ law.cfg
@@ -20,12 +20,12 @@ gfal2: WARNING
 [analysis]
 
 default_analysis: hbw.config.analysis_hbw.analysis_hbw
-default_config: config_2017_limited
+default_config: c17
 default_dataset: ggHH_kl_1_kt_1_sl_hbbhww_powheg
 
 production_modules: columnflow.production.{categories,processes,pileup,normalization,seeds}, hbw.production.{weights,features,ml_inputs,categories,gen_hbw_decay}
 calibration_modules: columnflow.calibration.jets, hbw.calibration.default
-selection_modules: hbw.selection.{categories,default}
+selection_modules: hbw.selection.{categories,common,sl,dl}
 ml_modules: hbw.ml.{derived,base,dense_classifier}
 inference_modules: hbw.inference.{test,default,new_model}
